+++ pkg/ingester/ingester_v2.go
@@ -24,6 +24,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/common/model"
+	promcfg "github.com/prometheus/prometheus/config"
 	"github.com/prometheus/prometheus/pkg/exemplar"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
@@ -676,6 +677,9 @@ func (i *Ingester) updateLoop(ctx context.Context) error {
 	ingestionRateTicker := time.NewTicker(instanceIngestionRateTickInterval)
 	defer ingestionRateTicker.Stop()
 
+	exemplarUpdateTicker := time.NewTicker(i.cfg.ExemplarsUpdatePeriod)
+	defer exemplarUpdateTicker.Stop()
+
 	var activeSeriesTickerChan <-chan time.Time
 	if i.cfg.ActiveSeriesMetricsEnabled {
 		t := time.NewTicker(i.cfg.ActiveSeriesMetricsUpdatePeriod)
@@ -700,6 +704,9 @@ func (i *Ingester) updateLoop(ctx context.Context) error {
 				db.ingestedRuleSamples.Tick()
 			}
 			i.userStatesMtx.RUnlock()
+
+		case <-exemplarUpdateTicker.C:
+			i.applyExemplarsSettings()
 
 		case <-activeSeriesTickerChan:
 			i.v2UpdateActiveSeries()
@@ -726,6 +733,33 @@ func (i *Ingester) v2UpdateActiveSeries() {
 		i.metrics.activeSeriesPerUser.WithLabelValues(userID).Set(float64(allActive))
 		for idx, name := range i.activeSeriesMatcher.MatcherNames() {
 			i.metrics.activeSeriesCustomTrackersPerUser.WithLabelValues(userID, name).Set(float64(activeMatching[idx]))
+		}
+	}
+}
+
+// Go through all tenants and apply the current max-exemplars setting.
+// If it changed, tsdb will resize the buffer; if it didn't change tsdb will return quickly.
+func (i *Ingester) applyExemplarsSettings() {
+	for _, userID := range i.getTSDBUsers() {
+		globalValue := i.limits.MaxGlobalExemplarsPerUser(userID)
+		localValue := i.limiter.convertGlobalToLocalLimit(userID, globalValue)
+		// We populate a Config struct with just one value, which is OK
+		// because Head.ApplyConfig only looks at one value.
+		// The other fields in Config are things like Rules, Scrape
+		// settings, which don't apply to Head.
+		cfg := promcfg.Config{
+			StorageConfig: promcfg.StorageConfig{
+				ExemplarsConfig: &promcfg.ExemplarsConfig{
+					MaxExemplars: int64(localValue),
+				},
+			},
+		}
+		tsdb := i.getTSDB(userID)
+		if tsdb == nil {
+			continue
+		}
+		if err := tsdb.Head().ApplyConfig(&cfg); err != nil {
+			level.Error(i.logger).Log("msg", "failed to apply config to TSDB", "user", userID, "err", err)
 		}
 	}
 }
@@ -882,10 +916,10 @@ func (i *Ingester) v2Push(ctx context.Context, req *mimirpb.WriteRequest) (*mimi
 			})
 		}
 
-		if i.cfg.BlocksStorageConfig.TSDB.MaxExemplars > 0 {
+		if len(ts.Exemplars) > 0 && i.limits.MaxGlobalExemplarsPerUser(userID) > 0 {
 			// app.AppendExemplar currently doesn't create the series, it must
 			// already exist.  If it does not then drop.
-			if ref == 0 && len(ts.Exemplars) > 0 {
+			if ref == 0 {
 				updateFirstPartial(func() error {
 					return wrappedTSDBIngestExemplarErr(errExemplarRef,
 						model.Time(ts.Exemplars[0].TimestampMs), ts.Labels, ts.Exemplars[0].Labels)
@@ -1642,10 +1676,7 @@ func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 		instanceSeriesCount: &i.TSDBState.seriesCount,
 	}
 
-	enableExemplars := false
-	if i.cfg.BlocksStorageConfig.TSDB.MaxExemplars > 0 {
-		enableExemplars = true
-	}
+	maxExemplars := i.limiter.convertGlobalToLocalLimit(userID, i.limits.MaxGlobalExemplarsPerUser(userID))
 	// Create a new user database
 	db, err := tsdb.Open(udir, userLogger, tsdbPromReg, &tsdb.Options{
 		RetentionDuration:         i.cfg.BlocksStorageConfig.TSDB.Retention.Milliseconds(),
@@ -1658,8 +1689,8 @@ func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 		WALSegmentSize:            i.cfg.BlocksStorageConfig.TSDB.WALSegmentSizeBytes,
 		SeriesLifecycleCallback:   userDB,
 		BlocksToDelete:            userDB.blocksToDelete,
-		EnableExemplarStorage:     enableExemplars,
-		MaxExemplars:              int64(i.cfg.BlocksStorageConfig.TSDB.MaxExemplars),
+		EnableExemplarStorage:     true, // enable for everyone so we can raise the limit later
+		MaxExemplars:              int64(maxExemplars),
 		SeriesHashCache:           i.TSDBState.seriesHashCache,
 	}, nil)
 	if err != nil {